+++ libs/vscode/shared/src/index.ts
@@ -2,5 +2,6 @@ export * from './lib/get-active-pro-code-or-task-window';
 export * from './lib/get-active-pro-code-window';
 export * from './lib/idl-command-action.interface';
 export * from './lib/open-file-in-vscode';
+export * from './lib/variables-reference-subsitution';
 export * from './lib/vscode-commands.interface';
 export * from './lib/vscode-telemetry-logger';